Final changes done

+++ OnlineBikeStore/Controllers/StockController.cs
@@ -1,11 +1,7 @@
-﻿using AutoMapper;
-using OnlineBikeStore.Models;
+﻿using OnlineBikeStore.Models;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Web;
 using System.Web.Mvc;
-using System.Data.Entity;
 namespace OnlineBikeStore.Controllers
 {
     public class StockController : Controller
@@ -20,7 +16,7 @@ namespace OnlineBikeStore.Controllers
         {
             return View();
         }
-       
+
 
         [HttpGet]
         public JsonResult GetProductsNotInAnyStore()
@@ -28,27 +24,44 @@ namespace OnlineBikeStore.Controllers
             var productsNotInAnyStore = context.products
                         .Where(p => !context.stocks
                         .Any(s => s.product_id == p.product_id))
-                        .Select(x=>new ProductViewModel
+                        .Select(x => new ProductViewModel
                         {
                             product_id = x.product_id,
                             product_name = x.product_name,
                             model_year = x.model_year
                         })
-                        .ToList();          
+                        .ToList();
 
             return Json(productsNotInAnyStore, JsonRequestBehavior.AllowGet);
         }
 
         [HttpPost]
-        public JsonResult AddStock(StockViewModel data)
+        public JsonResult AddOrUpdateStock(StockViewModel data)
         {
             try
             {
-                stock s = new stock();
-                s.product_id = data.product_id;
-                s.store_id = data.store_id;
-                s.quantity = data.quantity;
-                var newStock = context.stocks.Add(s);
+                // Check if the stock already exists for the given product and store
+                var existingStock = context.stocks
+                    .FirstOrDefault(s => s.product_id == data.product_id && s.store_id == data.store_id);
+
+                if (existingStock != null)
+                {
+                    // Update the existing stock quantity
+                    existingStock.quantity = data.quantity;
+                }
+                else
+                {
+                    // Add new stock
+                    stock newStock = new stock
+                    {
+                        product_id = data.product_id,
+                        store_id = data.store_id,
+                        quantity = data.quantity
+                    };
+                    context.stocks.Add(newStock);
+                }
+
+                // Save changes to the database
                 context.SaveChanges();
                 return Json(new { success = true });
             }
@@ -57,16 +70,29 @@ namespace OnlineBikeStore.Controllers
                 return Json(new { success = false, message = ex.Message });
             }
         }
+
         [HttpPost]
-        public JsonResult RemoveStock(StockViewModel data)
+        public JsonResult RemoveStockFromStore(int product_id, int store_id)
         {
             try
             {
-                var stock = context.stocks.Where(s => s.product_id == data.product_id && s.store_id == data.store_id).FirstOrDefault();
+                if (product_id < 1 || store_id < 1)
+                {
+                    return Json(new { success = false, message = "Invalid data" });
+                }
+
+                var stock = context.stocks
+                    .Where(s => s.product_id == product_id && s.store_id == store_id)
+                    .FirstOrDefault();
+
+                if (stock == null)
+                {
+                    return Json(new { success = false, message = "Stock not found" });
+                }
 
                 context.stocks.Remove(stock);
                 context.SaveChanges();
-                return Json(new { success = true });
+                return Json(new { success = true, message = "Item Reomved" });
             }
             catch (Exception ex)
             {
